Se añadio codigo en el JS de cada interfaz para que no puedan acceder a otras interfaces al menos que esten en una sesion
