+++ login_usuarios/login_usuario.html
@@ -61,6 +61,7 @@
             <input type="password" id="contraseña" name="contraseña" required>
             
             <button type="submit">Ingresar</button>
+            <p class="inicio">Aun no tienes una cuenta? <a href="../signup_clientes/signUpClientes.html">Registrate</a></p>
         </form>
         <div id="info-message" class="info-message">Iniciando sesión...</div>
         <div id="error-message" class="error-message"></div>
